report refactoring, added fam report
