+++ src/etools/applications/management/urls.py
@@ -1,10 +1,11 @@
-
 from django.conf.urls import url
 
 from etools.applications.management.views.general import InvalidateCache, SyncCountries, SyncFRs
 from etools.applications.management.views.reports import LoadResultStructure
 from etools.applications.management.views.tasks_endpoints import (
+    FamReportView,
     PMPIndicatorsReportView,
+    SyncDeltaUsers,
     TestSendEmailAPIView,
     UpdateAggregateHactValuesAPIView,
     UpdateHactValuesAPIView,
@@ -23,10 +24,12 @@ urlpatterns = ((
     url(r'^sync-countries/$', SyncCountries.as_view(), name='sync_frs'),
     url(r'^api/stats/usercounts/$', ActiveUsers.as_view(), name='stats_user_counts'),
     url(r'^api/stats/agreements/$', AgreementsStatisticsView.as_view(), name='stats_agreements'),
+    url(r'^tasks/sync_delta_users/$', SyncDeltaUsers.as_view(), name='tasks_sync_delta_users'),
     url(r'^tasks/update_hact_values/$', UpdateHactValuesAPIView.as_view(), name='tasks_update_hact_values'),
     url(r'^tasks/update_aggregate_hact_values/$', UpdateAggregateHactValuesAPIView.as_view(),
         name='tasks_aggregate_update_hact_values'),
     url(r'^tasks/send_test_email/$', TestSendEmailAPIView.as_view(), name='tasks_send_test_email'),
     url(r'^reports/users/$', UsersReportView.as_view(), name='reports_users'),
     url(r'^reports/pmp_indicators/$', PMPIndicatorsReportView.as_view(), name='reports_pmp_indicators'),
+    url(r'^reports/fam/$', FamReportView.as_view(), name='reports_fam'),
 ), 'management')
